Screenshot offscreen-render fix + shared tab-control/pane refactor
Offscreen screenshots (Scope / FFT / FreqResp):
- Fix the headless tab-strip clip: the screenshot clone was collapsed
  before its first expanded layout, so the CTabFolder reported a stale,
  too-short strip height and printed the bottom tiles / tab outlines /
  descenders clipped. renderOffscreen now lays the clone out expanded,
  lets it settle, then collapses - mirroring the live pane.
- Drop the redundant manual tile overlay: Control.print() already
  captures the folder's tile paint listener, so paintSnapshotTiles ->
  paintTilesInto double-drew every tile. Tile rendering is now fully
  encapsulated in TileTabFolder's own paint listener.
- De-duplicate: extract groupCell into AbstractTabControl and
  setTabsCollapsed into AbstractPane (via a tabStrip() getter); drop the
  per-pane copies. Move ScreenshotDialog / ScreenshotOverlay to gui.common.

Also folds in accumulated tree work: mains-suppression filter rework
(MainsCombFilter + new MainsFilters / MainsFrequencyTracker /
Mains{Lms,SyncSubtract,Time}Filter), StereoSampleFloat -> StereoSampleDouble,
predistortion engine/wizard updates, PresetBar widget, theory help
articles, and i18n keys across all locales.

Build + 256 tests green.

+++ src/main/java/org/edgo/audio/measure/cli/RegressCalibrateMode.java
@@ -71,7 +71,7 @@ public class RegressCalibrateMode {
         log.info("Scale     : {} V", scaleVolts);
         log.info("Chart     : {}x{} px", chartWidth, chartHeight);
 
-        float[] samples   = new float[(int) Math.min(totalFrames, Integer.MAX_VALUE)];
+        double[] samples   = new double[(int) Math.min(totalFrames, Integer.MAX_VALUE)];
         long    halfRange = 1L << (bitDepth - 1);
         AtomicInteger pos = new AtomicInteger(0);
 
@@ -79,13 +79,13 @@ public class RegressCalibrateMode {
             int n = pos.get();
             for (int i = 0; i < block.length && n + i < samples.length; i++) {
                 int signed = block[i].ch1 - (int) halfRange;
-                samples[n + i] = (float) (signed / (double) halfRange);
+                samples[n + i] = signed / (double) halfRange;
             }
             pos.addAndGet(block.length);
         });
 
         int actualSamples = pos.get();
-        float[] trimmed = actualSamples < samples.length
+        double[] trimmed = actualSamples < samples.length
                 ? Arrays.copyOf(samples, actualSamples)
                 : samples;
